Implement slash command registering
for the new library version.
